transaction/requirements.txt
fastapi
redis_om
python-dotenv
uvicorn
python-multipart
requests
python-jose

